+++ sdk/identity/identity/CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Bugs Fixed
 
 - Fixed bug that duplicated the tenant Id on the URI of outgoing requests when passing an `authorityHost` ending with a tenant Id.
+- `ManagedIdentityCredential` now won't retry when it tries to ping the IMDS endpoint.
+- Now we are specifying the maximum number of retries to 3 to ensure that maximum retries won't change without notice.
 
 ## 2.0.2 (2022-02-03)